requirements.txt
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
matplotlib>=3.7
statsmodels>=0.14
scikit-learn>=1.3
PyPortfolioOpt>=1.5
cvxpy>=1.4
seaborn>=0.13
pyyaml>=6.0
streamlit>=1.33
